feat: add RedeemUI
